requirements.txt
langchain
chromadb
openai
unstructured
pillow
